requirements.txt
python-statemachine==2.5.0